+++ lib/dialogs/dialogs.dart
@@ -1,0 +1,3 @@
+library dialogs;
+
+export 'confirmation_dialog.dart';