+++ lib/extensions/extensions.dart
@@ -2,3 +2,4 @@ library extensions;
 
 export 'list_extension.dart';
 export 'string_extension.dart';
+export 'widget_extension.dart';