+++ lib/extensions/string_extension.dart
@@ -1,9 +1,25 @@
 extension StringExtension on String {
   /// Capitalize the first letter of a word within a string
   /// Examples:
-  /// test this beautiful text -> Test this beautiful text
-  /// Test This Beautiful Text -> Test this beautiful text
+  /// "test this beautiful text" -> "Test this beautiful text"
+  /// "Test This Beautiful Text" -> "Test this beautiful text"
   String capitalize() {
     return "${this[0].toUpperCase()}${substring(1).toLowerCase()}";
   }
+
+  /// Checks wether the string does not contains characters
+  /// Examples:
+  /// "This is a string" -> false
+  /// " This is a string " -> false
+  /// "" -> true
+  /// "   " -> true
+  bool isBlank() => trim().isEmpty;
+
+  /// Checks wether the string contains characters
+  /// Examples:
+  /// "This is a string" -> true
+  /// " This is a string " -> true
+  /// "" -> false
+  /// "   " -> false
+  bool isNotBlank() => !isBlank();
 }