+++ lib/extensions/widget_extension.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/widgets.dart';
+
+extension WidgetExtension on Widget {
+  Widget center() {
+    /// Center any widget
+    return Center(
+      child: this,
+    );
+  }
+}