+++ lib/widgets/widgets.dart
@@ -1,3 +1,4 @@
 library widgets;
 
 export 'input_box.dart';
+export 'message.dart';